requirements.txt
# Meta Marketing API
facebook-business==19.0.0

# Notion API
notion-client==2.2.1

# Data Processing
pandas>=2.0.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests>=2.31.0

# Logging
colorlog>=6.7.0
